requirements.txt
asgiref==3.8.1
Django==4.2
django-cleanup==9.0.0
django-jalali==7.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-widget-tweaks==1.5.0
jalali_core==1.0.0
jdatetime==5.0.0
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
